comments and moderationCount fixed, minor fixes in postImage

+++ DevBlogProject/src/main/java/main/api/response/dto/AuthCheckUserDTO.java
@@ -14,7 +14,7 @@ public class AuthCheckUserDTO {
 
   private boolean moderation;
 
-  private int moderationCount;
+  private long moderationCount;
 
   private boolean settings;
 }
